walltime bugfix

+++ src/save_info.h
@@ -1,4 +1,4 @@
 #ifndef SAVE_INFO_H
 #define SAVE_INFO_H
-int save_info(const char *filename, double elapsedTime, const char *method, int m, int k, int n, int nproc, int nthreads, int total, int seqt);
+int save_info(const char *filename, double elapsedTime, const char *method, int m, int k, int n, int nproc, int nthreads, double total, double seqt);
 #endif
